refactor(phase-3-2): split theme_optimizer.py into package (503→4 files)
拆分方案:
- _presets.py (236): ThemePresets 类 (DARK/LIGHT 主题配色)
- _selector.py (89):  ThemePresetSelector UI (主题选择器)
- _preview.py (51):  ThemeColorPreview UI (颜色预览)
- _styles.py (143):  generate_theme_stylesheet (样式表生成)
- __init__.py (21): 重导出 4 个公共符号

兼容:
- from scenefab.ui.theme.theme_optimizer import X 仍工作 (通过 __init__.py)
- 现有 1 个引用点 (theme_toggle.py) 不需修改

验证:
- ruff check all green
- pytest 351 passed, 20 skipped
- F821 修复: _selector.py 添加 ThemePresets import
