+++ src/scenefab/ui/theme/theme_optimizer/_presets.py
@@ -1,0 +1,236 @@
+from PySide6.QtCore import Qt, Signal
+from PySide6.QtGui import QColor, QFont, QPainter
+from PySide6.QtWidgets import QHBoxLayout, QLabel, QPushButton, QVBoxLayout, QWidget
+
+from scenefab.ui.components.design_system import Colors
+
+
+class ThemePresets:
+    """主题配色方案预设"""
+
+    # 深色主题系列
+    DARK_THEMES = {
+        "Midnight": {
+            "primary": "#6366F1",
+            "primary_end": "#8B5CF6",
+            "accent": "#06B6D4",
+            "background": "#0A0A0F",
+            "surface": "#12121A",
+            "card": "#1A1A24",
+            "card_elevated": "#22222E",
+            "text": "#E6EDF3",
+            "text_secondary": "#C9D1D9",
+            "text_tertiary": "#8B949E",
+            "border": "#30363D",
+            "divider": "#21262D",
+            "success": "#238636",
+            "warning": "#D29922",
+            "error": "#DA3633",
+            "info": "#388BFD",
+        },
+        "Ocean": {
+            "primary": "#0EA5E9",
+            "primary_end": "#06B6D4",
+            "accent": "#14B8A6",
+            "background": "#0C1929",
+            "surface": "#132F4C",
+            "card": "#173A5E",
+            "card_elevated": "#1E4976",
+            "text": "#E3F2FD",
+            "text_secondary": "#90CAF9",
+            "text_tertiary": "#64B5F6",
+            "border": "#2196F3",
+            "divider": "#1565C0",
+            "success": "#2E7D32",
+            "warning": "#F57C00",
+            "error": "#D32F2F",
+            "info": "#1976D2",
+        },
+        "Forest": {
+            "primary": "#22C55E",
+            "primary_end": "#16A34A",
+            "accent": "#14B8A6",
+            "background": "#052E16",
+            "surface": "#064E3B",
+            "card": "#14532D",
+            "card_elevated": "#166534",
+            "text": "#ECFDF5",
+            "text_secondary": "#A7F3D0",
+            "text_tertiary": "#6EE7B7",
+            "border": "#16A34A",
+            "divider": "#15803D",
+            "success": "#22C55E",
+            "warning": "#F59E0B",
+            "error": "#EF4444",
+            "info": "#3B82F6",
+        },
+        "Sunset": {
+            "primary": "#F97316",
+            "primary_end": "#EF4444",
+            "accent": "#F43F5E",
+            "background": "#1A0A0A",
+            "surface": "#2C1010",
+            "card": "#3D1515",
+            "card_elevated": "#4E1A1A",
+            "text": "#FEE2E2",
+            "text_secondary": "#FCA5A5",
+            "text_tertiary": "#F87171",
+            "border": "#DC2626",
+            "divider": "#B91C1C",
+            "success": "#22C55E",
+            "warning": "#F59E0B",
+            "error": "#EF4444",
+            "info": "#3B82F6",
+        },
+        "Purple": {
+            "primary": "#A855F7",
+            "primary_end": "#7C3AED",
+            "accent": "#EC4899",
+            "background": "#1E1B2E",
+            "surface": "#2D2A42",
+            "card": "#3D3856",
+            "card_elevated": "#4D456A",
+            "text": "#F5F3FF",
+            "text_secondary": "#C4B5FD",
+            "text_tertiary": "#A78BFA",
+            "border": "#7C3AED",
+            "divider": "#6D28D9",
+            "success": "#22C55E",
+            "warning": "#F59E0B",
+            "error": "#EF4444",
+            "info": "#3B82F6",
+        },
+        "Rose": {
+            "primary": "#F43F5E",
+            "primary_end": "#E11D48",
+            "accent": "#FB7185",
+            "background": "#1A0F13",
+            "surface": "#2B131A",
+            "card": "#3C1821",
+            "card_elevated": "#4D1D28",
+            "text": "#FFE4E6",
+            "text_secondary": "#FDA4AF",
+            "text_tertiary": "#FB7185",
+            "border": "#BE123C",
+            "divider": "#9F1239",
+            "success": "#22C55E",
+            "warning": "#F59E0B",
+            "error": "#F43F5E",
+            "info": "#3B82F6",
+        },
+    }
+
+    # 浅色主题系列
+    LIGHT_THEMES = {
+        "Snow": {
+            "primary": "#0969DA",
+            "primary_end": "#8250DF",
+            "accent": "#BF3989",
+            "background": "#FFFFFF",
+            "surface": "#F6F8FA",
+            "card": "#FFFFFF",
+            "card_elevated": "#FFFFFF",
+            "text": "#1F2328",
+            "text_secondary": "#656D76",
+            "text_tertiary": "#8C959F",
+            "border": "#D0D7DE",
+            "divider": "#D8DEE4",
+            "success": "#1A7F37",
+            "warning": "#9A6700",
+            "error": "#CF222E",
+            "info": "#0969DA",
+        },
+        "Mint": {
+            "primary": "#059669",
+            "primary_end": "#0891B2",
+            "accent": "#0D9488",
+            "background": "#ECFDF5",
+            "surface": "#D1FAE5",
+            "card": "#FFFFFF",
+            "card_elevated": "#FFFFFF",
+            "text": "#022C22",
+            "text_secondary": "#064E3B",
+            "text_tertiary": "#047857",
+            "border": "#A7F3D0",
+            "divider": "#D1FAE5",
+            "success": "#059669",
+            "warning": "#D97706",
+            "error": "#DC2626",
+            "info": "#0284C7",
+        },
+        "Sky": {
+            "primary": "#0284C7",
+            "primary_end": "#6366F1",
+            "accent": "#8B5CF6",
+            "background": "#F0F9FF",
+            "surface": "#E0F2FE",
+            "card": "#FFFFFF",
+            "card_elevated": "#FFFFFF",
+            "text": "#0C4A6E",
+            "text_secondary": "#075985",
+            "text_tertiary": "#0369A1",
+            "border": "#BAE6FD",
+            "divider": "#E0F2FE",
+            "success": "#059669",
+            "warning": "#D97706",
+            "error": "#DC2626",
+            "info": "#0284C7",
+        },
+        "Peach": {
+            "primary": "#EA580C",
+            "primary_end": "#DC2626",
+            "accent": "#F43F5E",
+            "background": "#FFF7ED",
+            "surface": "#FFEDD5",
+            "card": "#FFFFFF",
+            "card_elevated": "#FFFFFF",
+            "text": "#7C2D12",
+            "text_secondary": "#9A3412",
+            "text_tertiary": "#C2410C",
+            "border": "#FED7AA",
+            "divider": "#FFEDD5",
+            "success": "#059669",
+            "warning": "#D97706",
+            "error": "#DC2626",
+            "info": "#0284C7",
+        },
+        "Lavender": {
+            "primary": "#7C3AED",
+            "primary_end": "#A855F7",
+            "accent": "#EC4899",
+            "background": "#FAF5FF",
+            "surface": "#F3E8FF",
+            "card": "#FFFFFF",
+            "card_elevated": "#FFFFFF",
+            "text": "#5B21B6",
+            "text_secondary": "#6D28D9",
+            "text_tertiary": "#7C3AED",
+            "border": "#E9D5FF",
+            "divider": "#F3E8FF",
+            "success": "#059669",
+            "warning": "#D97706",
+            "error": "#DC2626",
+            "info": "#0284C7",
+        },
+    }
+
+    @classmethod
+    def get_all_themes(cls):
+        """获取所有主题"""
+        return {
+            "深色主题": cls.DARK_THEMES,
+            "浅色主题": cls.LIGHT_THEMES,
+        }
+
+    @classmethod
+    def get_theme_names(cls, theme_type: str = None):
+        """获取主题名称列表"""
+        _THEME_MAP = {
+            "dark": cls.DARK_THEMES,
+            "light": cls.LIGHT_THEMES,
+        }
+        if theme_type in _THEME_MAP:
+            return list(_THEME_MAP[theme_type].keys())
+        return list(cls.DARK_THEMES.keys()) + list(cls.LIGHT_THEMES.keys())
+
+
